requirements.txt
aiohttp==3.9.5
aiosignal==1.3.2
alembic==1.14.1
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
asgiref==3.8.1
asttokens==3.0.0
attrs==23.2.0
Authlib==1.3.1
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.2.0b12
azure-mgmt-resource==23.1.0b2
azure-mgmt-subscription==3.0.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.0.1
Beaker==1.12.1
beautifulsoup4==4.12.3
bidict==0.22.1
blinker==1.7.0
blivet==3.10.0
blivet-gui==2.6.0
blosc2==2.5.1
boto3==1.37.38
botocore==1.37.38
Bottleneck==1.3.7
branca==0.8.1
Brlapi==0.8.5
Brotli==1.1.0
cachetools==5.3.2
certifi==2025.4.26
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
contourpy==1.2.0
cryptography==41.0.7
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
dasbus==1.7
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
distlib==0.4.0
distro==1.9.0
Django==5.2.4
dnf==4.23.0
dnspython==2.6.1
eduvpn_client==4.4.0
eduvpn_common==2.1.0
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.2.0
fb-re2==1.0.7
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.18.0
flaredantic==0.1.4
Flask==3.0.3
flask-babel==4.0.0
Flask-Compress==1.14
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.3.3
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
folium==0.19.5
fonttools==4.50.0
fros==1.1
frozenlist==1.4.1
fs==2.4.16
fsspec==2025.3.2
future==1.0.0
gcsfs==2023.6.0+1.g7cc53d9
geocoder==1.38.1
google-api-core==2.11.1
google-api-python-client==2.120.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.8.0
google-cloud-core==2.3.3
google-cloud-storage==2.14.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.63.0
greenlet==3.2.2
grpcio==1.48.4
grpcio-status==1.48.4
gssapi==1.7.3
h11==0.16.0
hf-xet==1.1.5
html5lib==1.1
httpagentparser==1.9.5
httpcore==1.0.9
httplib2==0.21.0
httpx==0.28.1
huggingface-hub==0.33.4
humanize==3.13.1
i3ipc==2.2.1
idna==3.7
iniconfig==1.1.1
instaloader==4.14.2
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
iso639==0.1.4
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jcopdl==2.3.3
jcopml==1.2.7
jdcal==1.4.1
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonformatter==0.3.2
jupyterlab_widgets==3.0.15
keyring==24.3.1
kiwisolver==1.4.8
langtable==0.0.68
ldap3==2.9.1
libcomps==0.1.20
libdnf==0.74.0
libgravatar==1.0.4
louis==3.28.0
luwiji==1.2.6
lxml==5.1.0
Mako==1.2.3
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mercurial==6.7.4
meson==1.4.1
more-itertools==10.1.0
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.2.0
msgpack==1.0.6
msrest==0.7.1
multidict==6.0.5
mutagen==1.47.0
ndindex==1.7
neo4j==5.28.1
neomodel==5.5.0
networkx==3.5
nftables==0.1
numexpr==2.8.5
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nwg-displays==0.3.25
oauthlib==3.2.2
odfpy==1.4.1
olefile==0.47
opencage==3.1.0
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
paramiko==3.5.1
parso==0.8.4
passlib==1.7.4
Paste==3.7.1
patsy==1.0.1
pexpect==4.9.0
phonenumbers==9.0.5
pid==2.2.3
pillow==10.3.0
pip-autoremove==0.10.0
platformdirs==3.11.0
playwright==1.52.0
pluggy==1.3.0
ply==3.11
pooch==1.8.2
portalocker==2.8.2
productmd==1.45
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.19.6
psutil==5.9.8
psycopg==3.1.18
psycopg2==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pwquality==1.4.5
py-cpuinfo==9.0.0
pyaml==25.7.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.25.1
pycares==4.3.0
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.21.0
pycups==2.0.4
pyee==13.0.0
pyenchant==3.2.2
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.6.0
pykickstart==3.52
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyotp==2.9.0
pyparsing==3.1.2
pyparted==3.13.0
pyperclip==1.9.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyquery==1.4.3
PySocks==1.7.1
pytest==7.4.3
python-augeas==1.1.0
python-dateutil==2.8.2
python-engineio==4.10.1
python-meh==0.51
python-pam==2.0.2
python-socketio==5.11.4
python-xlib==0.33
pytz==2025.2
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
qrcode==7.4.2
QtPy==2.4.1
rapidfuzz==3.5.2
ratelim==0.1.6
regex==2024.11.6
requests==2.31.0
requests-file==2.0.0
requests-ftp==0.3.1
requests-futures==1.0.0
requests-oauthlib==1.3.1
rich==13.7.0
rpm==4.19.1.1
rsa==4.9
s3transfer==0.11.5
safetensors==0.5.3
scikit-learn==1.7.0
scikit-optimize==0.10.2
SciPy==1.11.3
seaborn==0.13.2
SecretStorage==3.3.3
selinux @ file:///builddir/build/BUILD/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==79.0.1
sherlock-project==0.15.0
shtab==1.6.1
simple-websocket==1.0.0
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
snscrape==0.7.0.20230622
sos==4.7.2
soupsieve==2.5
SQLAlchemy==2.0.40
sqlparse==0.4.2
sshtunnel==0.4.0
stack-data==0.6.3
statsmodels==0.14.5
stem==1.8.2
sympy==1.14.0
systemd-python==235
tables==3.9.2
tabulate==0.9.0
Tempita==0.5.2
terminator==2.1.4
thefuzz==0.22.1
threadpoolctl==3.6.0
TikTokApi==7.1.0
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.1
typer==0.11.1
typing_extensions==4.12.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
virtualenv==20.34.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wsproto==1.2.0
WTForms==3.0.1
xarray==2023.8.0
xkbregistry==0.3
xlrd==2.0.1
XlsxWriter==3.2.3
xyzservices==2025.4.0
yarl==1.20.0
yt-dlp==2025.3.31
zstandard==0.22.0
